[47] post-merge fixes
