+++ src/routes/swagger.rs
@@ -9,13 +9,14 @@ use crate::setup::AppState;
 use crate::routes::attendee_routes;
 use crate::routes::debate_routes;
 use crate::routes::motion_routes;
-use crate::routes::roles;
+use crate::routes::roles_routes;
 use crate::routes::team_routes;
 use crate::routes::tournament_routes;
 use crate::tournament;
 use crate::tournament::attendee;
 use crate::tournament::debate;
 use crate::tournament::motion;
+use crate::tournament::roles;
 use crate::tournament::team;
 use crate::users::permissions;
 use crate::users::photourl;
@@ -69,10 +70,10 @@ pub fn route() -> Router<AppState> {
         user::get_user_by_id,
         user::patch_user_by_id,
         user::delete_user_by_id,
-        roles::create_user_roles,
-        roles::get_user_roles,
-        roles::patch_user_roles,
-        roles::delete_user_roles,
+        roles_routes::create_user_roles,
+        roles_routes::get_user_roles,
+        roles_routes::patch_user_roles,
+        roles_routes::delete_user_roles,
     ),
     components(schemas(
         version::VersionDetails,
